requirements.txt
streamlit>=1.37,<2
pymongo>=4.10,<5
langchain>=1.0.0,<2
langchain-openai>=1.0.0,<2
langchain-mongodb>=0.11.0,<0.12
python-dotenv>=1.0.1
pydantic>=2.8,<3
